% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_cases.R
\name{string_upper}
\alias{string_upper}
\title{Modify cases for strings}
\usage{
string_upper(s)
string_lower(s)
string_swapcase(s)
string_titlecase(s)
}
\arguments{
\item{s}{A string (character) vector.}
}
\value{
Character vector.
}
\description{
Modify the case of the string.
}
\details{
The \code{upper()} functions make all characters in a string vector in their capitalized case. 

The \code{string_lower()} function makes the same vector in their lower case. 

The \code{string_swapcase()} function switches the case for each character in a string vector.

The \code{string_titlecase()} function converts the first character of each word to uppercase.

The synonym pattern of these functions are \code{s_*()} and \code{*()} (replace asterisks with upper, lower, swapcase, or titlecase).
}
\examples{
string_upper(rownames(mtcars))
string_lower(rownames(mtcars))
string_swapcase(rownames(mtcars))

}
\seealso{
\url{https://github.com/robertschnitman/stringops}
}
