% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_chomp.R
\name{string_chomp}
\alias{string_chomp}
\title{Remove (empty) characters from a string as inspired by Ruby}
\usage{
string_chomp(s)
s_chomp(s)
chomp(s)
string_chop(s)
s_chop(s)
chop(s)
string_trim(s, which, whitespace)
s_trim(s, which, whitespace)
trim(s, which, whitespace)
string_cut(s, which)
s_cut(s, which)
cut(s, which)
}
\arguments{
\item{s}{A string (character) vector.}

\item{which}{Denotes how to trim/cut a vector (both, left only, or right only).}

\item{whitespace}{Denotes which whitespace characters to remove.}
}
\value{
Character vector.
}
\description{
Remove (empty) characters from a string as inspired by Ruby.
}
\details{
As inspired by Ruby, \code{string_chomp} removes all whitespace characters in a string. The function \code{string_chop} removes the last character from a string. The function \code{string_trim} acts the same as \code{trimws} but has a more readable source code.
}
\examples{
string_chomp(rownames(mtcars))
string_chop(rownames(mtcars))
string_trim("   s   ")
string_cut("cut this please", which = 'both')

}
\seealso{
\url{https://github.com/robertschnitman/stringops}
}
