% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_compare.R
\name{string_isblank}
\alias{string_isblank}
\title{Compare string vectors.}
\usage{
string_isblank(s)
s_isblank(s)
isblank(s)
is.blank(s)
%like%
string_isupper(s)
s_isupper(s)
isupper(s)
is.upper(s)
string_islower(s)
s_islower(s)
islower(s)
is.lower(s)
}
\arguments{
\item{s}{A string (character) vector.}
}
\value{
Boolean vector.
}
\description{
Compare string vectors.
}
\details{

}
\examples{
string_isblank("         ")
string_isblank("    string     ")

rownames(mtcars) \%like\% "^M"

string_isupper(c('TEST', 'test', 'tEsT'))
string_islower(c('TEST', 'test', 'tEsT'))

}
\seealso{
\url{https://github.com/robertschnitman/stringops}
}
