% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_concat.R
\name{string_concat}
\alias{string_concat}
\title{Concatenate strings together}
\usage{
%&%
string_prefix(s, prefix)
s_prefix(s, prefix)
prefix(s, prefix)
string_suffix(s, suffix)
s_suffix(s, suffix)
suffix(s, suffix)
}
\arguments{
\item{s}{A string (character) vector.}

\item{collapse}{A string by which to separate the elements of a vector.}
}
\value{

}
\description{
Concatenate strings together.
}
\examples{
"a" \%&\% "b"
string_prefix(rownames(mtcars), "A")
string_suffix(rownames(mtcars), "Z")

}
\seealso{
\url{https://github.com/robertschnitman/stringops}
}
