% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_misc.R
\name{string_dup}
\alias{string_dup}
\title{Miscellaneous string functions}
\usage{
string_dup(s, times)
string_insert(s, insert, position)
string_split(s, split, ...)
string_len(s)
string_reverse(s)
}
\description{
Miscellaneous string functions that don't quite fit into the other categories.
}
\details{
\code{string_dup()} acts the same as strrep(): it repeats a string n number of times. 
\code{string_insert()} inserts a string at a specified position. 
\code{string_len()} acts the same as nchar(): it counts the number of characters per vector element.
\code{string_reverse()} reverses the characters in a string.
\code{string_split() acts the same as strsplit(): it splits a string by a specified character.}
}
\examples{
string_dup(rownames(mtcars), 3)
string_insert("abcd", "Z", 3)
string_len(rownames(mtcars))
string_reverse(rownames(mtcars))
string_split(rownames(mtcars), " ")

}
\seealso{
\url{https://github.com/robertschnitman/stringops}
}
