% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_spot.R
\name{string_spot}
\alias{string_spot}
\title{Spot patterns in strings}
\usage{
string_spot(s, pattern)
s_spot(s, pattern)
spot(s, pattern)
string_spoti(s, pattern)
s_spoti(s, pattern)
spoti(s, pattern)
string_spotl(s, pattern)
s_spotl(s, pattern)
spotl(s, pattern)
string_spotm(s, pattern)
s_spotm(s, pattern)
spotm(s, pattern)
string_replace(s, search, replace)
s_replace(s, search, replace)
spot_replace(s, search, replace)
fr(s, search, replace)
search_replace(s, search, replace)
sr(s, search, replace)
string_remove(s, remove)
s_remove(s, remove)
string_countm(s, pattern, ...)
s_countm(s, pattern, ...)
countm(s, pattern, ...)
}
\arguments{
\item{s}{A string (character) vector.}

\item{pattern}{A regular expression pattern.}

\item{search}{A pattern to search.}

\item{replace}{The string to replace the searched pattern.}

\item{remove}{The string to remove.}

\item{...}{Parameters to pass to \code{gregexpr}}
}
\value{
Character vector for all functions except the \code{spotl()} and \code{countm()} functions, which produce Boolean and numeric vectors respectively.
}
\description{
Spot patterns in strings.
}
\details{
The function \code{string_spot} subsets a vector to the values matching a given pattern. Its synonyms are \code{s_spot} and \code{spot}.

The function \code{string_spoti} provides the indices where a pattern in a vector is found. Its synonyms are \code{s_spoti} and \code{spoti}.

The function \code{string_spotl} detects whether a pattern exists in a vector. Its synonyms are \code{s_spotl} and \code{spotl}.

The \code{spotm} functions spots pattern matches and returns NA if none are found.

The \code{*_replace} functions act the same as \code{gsub}.

The \code{*_remove} functions blank out a matching pattern.

The \code{string_countm} counts the number of matches in a string.
}
\examples{
rn <- rownames(mtcars)

string_spot(rn, "^M")
string_spoti(rn, "^M")
string_spotl(rn, "^M")
string_spotm(rn, "^M")

string_replace(rn, "^M", "Z")
string_remove(rn, "^M")
string_countm(rn, 'a')

}
\seealso{
\url{https://github.com/robertschnitman/stringops}
}
