% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_cases.R
\name{string_upper}
\alias{string_upper}
\title{Modify cases for strings.}
\usage{
string_upper(s)
s_upper(s) 
upper(s) 
string_lower(s)
s_lower(s)
lower(s)
string_swapcase(s)
s_swapcase(s)
swapcase(s)
string_titlecase(s)
s_titlecase(s)
titlecase(s)
}
\arguments{
\item{s}{A string (character) vector.}
}
\value{
Character vector.
}
\description{
Modify the case of the string.
}
\details{
The \code{upper} functions make all characters in a string vector in their capitalized case. The \code{lower} functions makes the same vector in their lower case. The \code{swapcase} functions switches the case for each character in a string vector. The \code{titlecase} functions converts the first character of each word to uppercase.
}
\examples{
upper(rownames(mtcars))
lower(rownames(mtcars))
swapcase(rownames(mtcars))

}
\seealso{
\url{https://github.com/robertschnitman/stringops}
}
