% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_compare.R
\name{string_isblank}
\alias{string_isblank}
\title{Compare string vectors.}
\usage{
%like%
string_isblank(s)
string_isupper(s)
string_islower(s)
}
\arguments{
\item{s}{A string (character) vector.}
}
\value{
Boolean vector.
}
\description{
Compare string vectors.
}
\details{
The function \code{string_isblank()} tests whether a vector consists of only blank characters. Synonyms are \code{s_isblank()}, \code{isblank()}, and \code{is.blank()}.

The \code{\%like\%} operator tests whether a pattern match is found in a string. 

The functions \code{string_isupper()} and \code{string_islower()} test whether each element in a string vector is all uppercase or lowercase, respectively. Synonyms are \code{s_fun()}, \code{isfun()}, and \code{is.fun()} (replace "fun" with "upper" or "lower").
}
\examples{
string_isblank("         ")
string_isblank("    string     ")

rownames(mtcars) \%like\% "^M"

string_isupper(c('TEST', 'test', 'tEsT'))
string_islower(c('TEST', 'test', 'tEsT'))

}
\seealso{
\url{https://github.com/robertschnitman/stringops}
}
