% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_concat.R
\name{string_concat}
\alias{string_concat}
\title{Concatenate strings together}
\usage{
%&%
string_concat(a, b)
string_prefix(s, prefix)
string_suffix(s, suffix)
string_flag(s, flag, width, which = c('left', 'right', 'both'))
string_flagv(s, ...)
}
\arguments{
\item{a}{A string (character) vector.}

\item{b}{A string (character) vector.}

\item{s}{A string (character) vector.}

\item{collapse}{A string by which to separate the elements of a vector.}

\item{flag}{Character or number to append to a string.}

\item{width}{Number of characters a string should be.}

\item{which}{Side to flag a string.}

\item{...}{Parameters passed to \code{string_flag()}.}
}
\value{
Character vector.
}
\description{
Concatenate strings together.
}
\details{
The \code{\%&\%} operator acts similar to BASIC's \code{&}, concatenating two elements together.

\code{string_prefix()} and its synonyms prefix a string to a vector, while \code{string_suffix()} and its synonyms suffix a string to a vector.

\code{string_flag()} is a scalar function that appends a character or number to a string if it does not meet a specified width.

\code{string_flagv()} is a vectorized version of \code{string_flag()}.

The synonym pattern of these functions are \code{s_*()} and \code{*()} (replace asterisks with prefix, suffix, and flag(v)).
}
\examples{
"a" \%&\% "b"
string_prefix(rownames(mtcars), "A")
string_suffix(rownames(mtcars), "Z")
string_flag('123456789', '0', 10)

}
\seealso{
\url{https://github.com/robertschnitman/stringops}
}
