% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_concat.R
\name{string_concat}
\alias{string_concat}
\title{Concatenate strings together}
\usage{
%&%
string_concat(a, b)
string_prefix(s, prefix)
s_prefix(s, prefix)
prefix(s, prefix)
string_suffix(s, suffix)
s_suffix(s, suffix)
suffix(s, suffix)
}
\arguments{
\item{a}{A string (character) vector.}

\item{b}{A string (character) vector.}

\item{s}{A string (character) vector.}

\item{collapse}{A string by which to separate the elements of a vector.}
}
\value{
Character vector.
}
\description{
Concatenate strings together.
}
\details{
The \code{\%&\%} operator acts similar to BASIC's \code{&}, concatenating two elements together.

\code{string_prefix} and its synonyms prefix a string to a vector, while \code{string_suffix} and its synonyms suffix a string to a vector.
}
\examples{
"a" \%&\% "b"
string_prefix(rownames(mtcars), "A")
string_suffix(rownames(mtcars), "Z")

}
\seealso{
\url{https://github.com/robertschnitman/stringops}
}
