% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_cull.R
\name{string_cull}
\alias{string_cull}
\title{Cull/extract patterns from strings}
\usage{
string_cull(s, pattern)
s_cull(s, pattern)
cull(s, pattern)
string_extract(s, pattern)
s_extract(s, pattern)
extract(s, pattern)
string_left(s, n)
s_left(s, n)
left(s, n)
string_right(s, n)
s_right(s, n)
right(s, n)
string_mid(s, start, stop)
s_mid(s, start, stop)
mid(s, start, stop)
}
\arguments{
\item{s}{A string (character) vector.}

\item{pattern}{A regular expression pattern.}

\item{n}{The number of characters.}

\item{start}{The position in a string at which to start.}

\item{stop}{The position in string at which to stop.}
}
\value{
Character vector.
}
\description{
Cull or extract patterns from strings.
}
\details{
The function \code{string_cull} (and its synonyms \code{s_cull}, \code{cull}, \code{string_extract}, \code{s_extract}, \code{extract}) culls (or extracts) a pattern from a string vector--if no pattern is found, \code{NA} is returned.

The functions \code{string_left}/\code{s_left}/\code{left} and \code{string_right}/\code{s_right}/\code{right} act the same as Excel's \code{LEFT} and \code{RIGHT} functions. The \code{mid} functions act the same as Excel's \code{MID} function.
}
\examples{
string_cull(rownames(mtcars), '^M.*')
string_left(rownames(mtcars), 3)
string_right(rownames(mtcars), 3)
string_mid(rownames(mtcars), 2, 4)

}
\seealso{
\url{https://github.com/robertschnitman/stringops}
}
