% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_join.R
\name{string_join}
\alias{string_join}
\title{Join elements into a single string}
\usage{
string_join(s, collapse)
}
\arguments{
\item{s}{A string (character) vector.}

\item{collapse}{A string by which to separate the elements of a vector.}
}
\value{
Character vector.
}
\description{
Join (collapse) elements into a single string or vector.
}
\details{
\code{string_join()} joins vector elements into a single string. Synonyms are \code{s_join()} and \code{join()}.
}
\examples{
string_join(rownames(mtcars), ", ")
}
\seealso{
\url{https://github.com/robertschnitman/stringops}
}
