% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_misc.R
\name{string_dup}
\alias{string_dup}
\title{Miscellaneous stirng functions}
\usage{
string_dup(s, times)
string_insert(s, insert, position)
string_split(s, split, ...)
string_len(s)
string_reverse(s)
}
\description{
Miscellaneous string functions due for a full explanation. Explore at your own risk!
}
\examples{
string_dup(rownames(mtcars), 3)
string_insert("abcd", "Z", 3)
string_len(rownames(mtcars))
string_reverse(rownames(mtcars))

}
\seealso{
\url{https://github.com/robertschnitman/stringops}
}
