% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string_spot.R
\name{string_spot}
\alias{string_spot}
\title{Spot patterns in strings}
\usage{
string_spot(s, pattern, invert = FALSE)
s_spot(s, pattern, invert = FALSE)
spot(s, pattern, invert = FALSE)
string_spoti(s, pattern, invert = FALSE)
s_spoti(s, pattern, invert = FALSE)
spoti(s, pattern, invert = FALSE)
string_spotl(s, pattern, invert = FALSE)
s_spotl(s, pattern, invert = FALSE)
spotl(s, pattern, invert = FALSE)
string_spotm(s, pattern, invert = FALSE)
s_spotm(s, pattern, invert = FALSE)
spotm(s, pattern, invert = FALSE)
string_replace(s, search, replace)
s_replace(s, search, replace)
find_replace(s, search, replace)
fr(s, search, replace)
search_replace(s, search, replace)
sr(s, search, replace)
string_remove(s, remove)
s_remove(s, remove)
string_countm(s, pattern, ...)
s_countm(s, pattern, ...)
countm(s, pattern, ...)
}
\arguments{
\item{s}{A string (character) vector.}

\item{pattern}{A regular expression pattern.}

\item{invert}{Boolean value (\code{TRUE/FALSE}) for whether to find actual matches (\code{FALSE}; default) or non-matches (\code{TRUE}).}

\item{search}{A pattern to search.}

\item{replace}{The string to replace the searched pattern.}

\item{remove}{The string to remove.}

\item{...}{Parameters to pass to \code{gregexpr()}.}
}
\value{
Character vector for all functions except the \code{spotl()}, \code{spoti()}, and \code{countm()} functions, which produce Boolean, numeric, and numeric vectors respectively.
}
\description{
Spot patterns in strings.
}
\details{
The function \code{string_spot()} subsets a vector to the values matching a given pattern. By setting \code{invert = TRUE}, the function will subset the vector for non-matches.

The function \code{string_spoti()} provides the indices where a pattern in a vector is found. By setting \code{invert = TRUE}, the function will find the indices of non-matches.

The function \code{string_spotl()} detects whether a pattern exists in a vector, outputting a Boolean value (\code{TRUE/FALSE}). Setting \code{invert = TRUE} is the same as \code{!grepl()}.

The \code{string_spotm()} function spots pattern matches and returns \code{NA} if none are found. Setting \code{invert = TRUE} will make the function find non-matches and return \code{NA} for matches.

The \code{string_replace()} function acts the same as \code{gsub()}.

The \code{string_remove()} function blanks out a matching pattern.

The \code{string_countm()} counts the number of matches in a string.
}
\examples{
rn <- rownames(mtcars)

string_spot(rn, "^M")
string_spot(rn, '^M', invert = TRUE)
string_spoti(rn, "^M")
string_spotl(rn, "^M")
string_spotm(rn, "^M")

string_replace(rn, "^M", "Z")
string_remove(rn, "^M")
string_countm(rn, 'a')


}
\seealso{
\url{https://github.com/robertschnitman/stringops}
}
